extracting the useNodeSelectionState hook into its own module
