+++ console/src/hooks/useSelectNodeHandler.ts
@@ -6,7 +6,7 @@ import { useFusionAccessTranslations } from "@/hooks/useFusionAccessTranslations
 import type { IoK8sApiCoreV1Node } from "@/models/kubernetes/1.30/types";
 import { useStoreContext } from "@/contexts/store/context";
 import { getDigest } from "@/utils/crypto/hash";
-import type { NodeSelectionActions } from "@/components/NodesSelectionTableRow";
+import type { NodeSelectionActions } from "@/hooks/useNodeSelectionState";
 import type { State, Actions } from "@/contexts/store/types";
 
 const [storageRoleLabelKey, storageRoleLabelValue] =
